refactor(currency): Symbol resolution uses `Group::find_map`
Deprecate `Group::maybe_visit`.
Move the `visit_any` into its own module.
