+++ protocol/packages/currencies/src/native/mod.rs
@@ -1,6 +1,8 @@
+use std::{borrow::Borrow, iter};
+
 use serde::{Deserialize, Serialize};
 
-use currency::{AnyVisitor, Matcher, MaybeAnyVisitResult, MemberOf};
+use currency::{AnyVisitor, FilterMapT, FindMapT, Matcher, MaybeAnyVisitResult, MemberOf};
 
 use crate::payment::Group as PaymentGroup;
 
@@ -38,6 +40,23 @@ impl currency::Group for Group {
         V: AnyVisitor<Self::TopG>,
     {
         currency::maybe_visit_member::<_, Nls, Self::TopG, _>(matcher, visitor)
+    }
+
+    fn filter_map<FilterMap, FilterMapRef>(
+        _f: FilterMapRef,
+    ) -> impl Iterator<Item = FilterMap::Outcome>
+    where
+        FilterMap: FilterMapT<Self>,
+        FilterMapRef: Borrow<FilterMap> + Clone,
+    {
+        iter::empty()
+    }
+
+    fn find_map<FindMap>(_v: FindMap) -> Result<FindMap::Outcome, FindMap>
+    where
+        FindMap: FindMapT<Self>,
+    {
+        todo!()
     }
 }
