+++ protocol/packages/currencies/src/payment/mod.rs
@@ -1,6 +1,8 @@
+use std::{borrow::Borrow, iter};
+
 use serde::{Deserialize, Serialize};
 
-use currency::{AnyVisitor, Matcher, MaybeAnyVisitResult, MemberOf};
+use currency::{AnyVisitor, FilterMapT, Matcher, MaybeAnyVisitResult, MemberOf};
 
 use crate::{lease::Group as LeaseGroup, lpn::Group as LpnGroup, native::Group as NativeGroup};
 
@@ -41,6 +43,23 @@ impl currency::Group for Group {
         V: AnyVisitor<Self::TopG>,
     {
         unreachable!()
+    }
+
+    fn filter_map<FilterMap, FilterMapRef>(
+        _f: FilterMapRef,
+    ) -> impl Iterator<Item = FilterMap::Outcome>
+    where
+        FilterMap: FilterMapT<Self>,
+        FilterMapRef: Borrow<FilterMap> + Clone,
+    {
+        iter::empty()
+    }
+
+    fn find_map<FindMap>(_v: FindMap) -> Result<FindMap::Outcome, FindMap>
+    where
+        FindMap: currency::FindMapT<Self>,
+    {
+        todo!()
     }
 }
